refactor: uprava slozek ui
