+++ ui/index.html
@@ -8,7 +8,8 @@
   </head>
   <body>
     <!-- <div id="preact-program"></div> -->
-    <div id="preact-obchod"></div>
+    <!-- <div id="preact-obchod"></div> -->
+    <div id="preact-obchod-nastaveni"></div>
     <script>
       window.GAMECON_KONSTANTY = {
         IS_DEV_SERVER: true,
