+++ ui/src/api/obchod/types.ts
