--- ui/src/pages/index.ts
@@ -1,8 +1,0 @@
-import { renderProgram } from "./program"
-import { renderObchod } from "./obchod";
-
-export const renderPages = () =>{
-  renderProgram();
-  renderObchod();
-}
-
