+++ ui/src/pages/obchodNastaveni/app.less
@@ -1,0 +1,8 @@
+
+.shop {
+  &--container {
+    width: 80vw;
+    height: 60vh;
+  }
+}
+
