+++ ui/src/pages/obchodNastaveni/index.tsx
@@ -1,0 +1,1 @@
+export { ObchodNastaveni } from './app'
